Replace old parser with new
