+++ crates/filecheck/Cargo.toml
@@ -12,6 +12,6 @@ publish = false
 filecheck = "0.5.0" # { path = "/Users/sean/src/filecheck" }
 sonatina-ir = { path = "../ir" }
 sonatina-codegen = { path = "../codegen" }
-sonatina-parser2 = { path = "../parser2" }
+sonatina-parser = { path = "../parser" }
 termcolor = "1.1.2"
 walkdir = "2"
