+++ crates/parser/test_files/syntax/func/empty.sntn
