+++ crates/parser/test_files/syntax/func/simple.sntn
