+++ crates/parser/test_files/syntax/module/simple.sntn
