--- crates/parser2/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "sonatina-parser2"
-version = "0.0.3-alpha"
-edition = "2021"
-authors = ["Sonatina Developers"]
-license = "Apache-2.0"
-readme = "../../README.md"
-homepage = "https://github.com/fe-lang/sonatina/tree/main/crates/parser"
-repository = "https://github.com/fe-lang/sonatina"
-description = "Parser for sonatina-ir text format"
-categories = ["compilers", "parser", "wasm"]
-keywords = ["compiler", "evm", "wasm", "smart-contract"]
-
-[dependencies]
-ir = { package = "sonatina-ir", path = "../ir", version = "0.0.3-alpha" }
-sonatina-triple = { path = "../triple", version = "0.0.3-alpha" }
-smallvec = "1.7.0"
-cranelift-entity = "0.104"
-pest = "2.7.10"
-pest_derive = "2.7.10"
-pest-ast = "0.3.4"
-from-pest = "0.3.2"
-smol_str = "0.2.2"
-hex = "0.4.3"
-num-traits = { version = "0.2.19", default-features = false }
-either = { version = "1.12.0", default-features = false }
-annotate-snippets = "0.11.4"
-
-[dev-dependencies]
-dir-test = { git = "https://github.com/sbillig/dir-test", rev = "c4115dd" }
-insta = { version = "1.38.0" }
-indenter = "0.3.3"
-ariadne = { version = "0.4.1", features = ["auto-color"] }
